Set up of main.py
the client_id and client_secret are in a file called .env however as they contain unique code I have selected to ignore those files to limit the access to only myself.

rest of the main.py has been set up

virtual environment installed and set up as well but have chosed to ignore for security purpose.

Will add the set up instructions to read.me so that anyone wanting to download/use the repo can still do so as well.