Get artist information
I am able to pass the relevant data to get_artist_information which is then returning true if the artist from the top 50 songs globally is classified as a rap/hip hop artist.

There is a issue in that the sample of songs in the top 50 which are rap/hip hop is very low, only about 7 songs are classified into the rap/ hip hop genre which is an issue as if there isnt a big enough sample size then the statistics become less trustworthy.

Possible solutions to a low sample size:
- Using multiple playlists
- Search for hip hop and rap genre and get the popular songs and their release dates, however this won't tell me when the song itself was popular.
- change the question and broaden the genres from hip hop, rap and include more genres